requirements.txt
asgiref==3.9.1
Django==5.2.5
django-ratelimit==4.1.0
djangorestframework==3.16.1
sqlparse==0.5.3
